vidpf: Add code for tree traversal with path checks
Add a method on `VidpfKey` for evaluating a share of the prefix tree
from a VIDPF key. This method transparently implements some the
functionality required for attributed-based metrics. In particular, it
computes the hash the onehot and path checks required for Mastic
robustness. (The user is expected to check that the weight is in range,
i.e., using an FLP.)