Prepare to add support for Prio3Histogram
Add a new mode of operation, "plain_metrics", that will just run plain
Prio3. The new mode is not fully implemented; this commit just does some
refactoring in preparation.

There are new VDAFs, each with a different report (share) structure.
Accammadating this requires some refactoring. In particular, the
add_keys() and add_flp_*() RPCs are merged into a single RPC,
add_report_shares(), which takes a sequence of ReportShare enums.

While at it, fix a bug in the weighted heavy hitters mode. (We were
computing the length of the FLP query randomness incorrectly.)